requirements.txt
django-crispy-forms==1.14.0
gunicorn==20.1.0
python-decouple==3.6
